Trigger 'cycle-slide-before' before adding slide
A new event, 'cycle-slide-before' is added to enable polyfills to run on slide data coming from progressive when loader is also enabled. This helps with slides containing <picture> or <img srcset/> tags to hold the image, since most browsers require polyfills for this to display properly.

Since loader requires img.complete to return TRUE (and uses the img.load() handler to trigger adding the new slide) running the polyfill AFTER the new slide is in the DOM but BEFORE the slide has been processed further is necessary. I've added the callback as early as possible in this scenario, and find that it could also be of use if any other preprocessing of the raw slide object is necessary.

This callback isn't necessary in the normal add() function which is not overridden by the loader module, but as it is a useful trigger point, I've added it there as well for consistency.